1 Translate the SessionTest class to English.
